question_project/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-django==4.8.0
gunicorn==21.2.0
pytest-factoryboy==2.5.1
model_bakery==1.19.2